Add files via upload
This Power BI project visualizes a survey conducted with workers in the programming and data sector, using an Excel table containing data on various aspects of their careers and work environments. The aim is to explore and present valuable insights interactively, focusing on key factors such as the difficulty of entering the sector, salary differences based on job titles, work-life balance, job satisfaction, and more.